+++ frontend/src/components/assessment-creation/ErrorToast.tsx
@@ -1,0 +1,17 @@
+import React from "react";
+import { Alert, AlertIcon } from "@chakra-ui/react";
+
+interface ErrorToastProps {
+  errorMessage: string;
+}
+
+const ErrorToast = ({ errorMessage }: ErrorToastProps): React.ReactElement => {
+  return (
+    <Alert status="error">
+      <AlertIcon />
+      {errorMessage}
+    </Alert>
+  );
+};
+
+export default ErrorToast;
