+++ frontend/src/components/pages/admin/CreateAssessmentPage.tsx
@@ -1,5 +1,4 @@
 /* eslint-disable react/jsx-props-no-spreading */
-
 import React, { useState } from "react";
 import countryList from "react-select-country-list";
 import { useForm, Controller } from "react-hook-form";
@@ -10,15 +9,15 @@ import {
   Input,
   VStack,
   RadioGroup,
-  Stack,
+  HStack,
   Radio,
   Box,
   Text,
   Button,
   FormErrorMessage,
-  Alert,
-  AlertIcon,
 } from "@chakra-ui/react";
+import gradeOptions from "../../../constants/CreateAssessmentConstants";
+import ErrorToast from "../../assessment-creation/ErrorToast";
 
 const CreateAssessmentPage = (): React.ReactElement => {
   const {
@@ -28,11 +27,6 @@ const CreateAssessmentPage = (): React.ReactElement => {
     control,
   } = useForm();
 
-  type FormInputs = {
-    assessmentName: string;
-  };
-
-  const fromErrorMsg = "Please resolve all issues before publishing or saving";
   const [validSubmit, setValidSubmit] = useState(true);
 
   const onSubmit = (data: any, e: any) => {
@@ -44,58 +38,17 @@ const CreateAssessmentPage = (): React.ReactElement => {
   };
 
   const countryOptions = React.useMemo(() => countryList().getData(), []);
-  const gradeOptions = [
-    {
-      label: "K",
-      value: "k",
-    },
-    {
-      label: "Grade 1",
-      value: "1",
-    },
-    {
-      label: "Grade 2",
-      value: "2",
-    },
-    {
-      label: "Grade 3",
-      value: "3",
-    },
-    {
-      label: "Grade 4",
-      value: "4",
-    },
-    {
-      label: "Grade 5",
-      value: "5",
-    },
-    {
-      label: "Grade 6",
-      value: "6",
-    },
-    {
-      label: "Grade 7",
-      value: "7",
-    },
-    {
-      label: "Grade 8",
-      value: "8",
-    },
-  ];
 
   return (
     <VStack align="left" width="75%" spacing={8}>
-      <Text textStyle="eyebrow">BASIC INFORMATION</Text>
+      <Text textStyle="eyebrow">Basic Information</Text>
 
       {!validSubmit && (
-        <Alert status="error">
-          <AlertIcon />
-          {fromErrorMsg}
-        </Alert>
+        <ErrorToast errorMessage="Please resolve all issues before publishing or saving" />
       )}
 
       <FormControl isRequired isInvalid={Boolean(errors.assessmentName)}>
-        <FormLabel color="grey.400"> Assessment Name </FormLabel>
+        <FormLabel color="grey.400">Assessment Name</FormLabel>
         <Input
           placeholder="e.g. Ontario Grade 5 Pre-Term Assessment"
           {...register("assessmentName", {
@@ -111,35 +64,18 @@ const CreateAssessmentPage = (): React.ReactElement => {
           name="grade"
           rules={{ required: "Please select a grade" }}
           render={({
-            field: { onChange, onBlur, value, name, ref },
+            field: { onChange, value, name },
             fieldState: { error },
           }) => (
             <FormControl isRequired isInvalid={Boolean(error)}>
-              <FormLabel color="grey.400"> Grade Level </FormLabel>
+              <FormLabel color="grey.400">Grade Level</FormLabel>
               <Select
                 name={name}
-                ref={ref}
                 onChange={onChange}
-                onBlur={onBlur}
                 value={value}
                 options={gradeOptions}
                 placeholder=""
-                chakraStyles={{
-                  dropdownIndicator: (provided) => ({
-                    ...provided,
-                    bg: "transparent",
-                    px: 2,
-                    cursor: "inherit",
-                  }),
-                  indicatorSeparator: (provided) => ({
-                    ...provided,
-                    display: "none",
-                  }),
-                  placeholder: (provided) => ({
-                    ...provided,
-                    color: "grey.300",
-                  }),
-                }}
+                useBasicStyles
               />
               <FormErrorMessage>{error?.message}</FormErrorMessage>
             </FormControl>
@@ -157,17 +93,17 @@ const CreateAssessmentPage = (): React.ReactElement => {
             fieldState: { error },
           }) => (
             <FormControl isRequired isInvalid={Boolean(error)}>
-              <FormLabel color="grey.400"> Type of Assessment </FormLabel>
+              <FormLabel color="grey.400">Type of Assessment</FormLabel>
               <RadioGroup
                 name={name}
                 ref={ref}
                 onChange={onChange}
                 value={value}
               >
-                <Stack direction="column">
+                <VStack align="left" spacing={0.5}>
                   <Radio value="beginning">Beginning of Grade</Radio>
                   <Radio value="end">End of Grade</Radio>
-                </Stack>
+                </VStack>
               </RadioGroup>
               <FormErrorMessage>{error?.message}</FormErrorMessage>
             </FormControl>
@@ -177,44 +113,26 @@ const CreateAssessmentPage = (): React.ReactElement => {
 
       <FormControl>
         <Text textStyle="subtitle2" mb="2">
-          {" "}
-          Curriculum{" "}
+          Curriculum
         </Text>
-        <Stack direction="row" width="100%">
+        <HStack width="100%">
           <Controller
             control={control}
             name="country"
             rules={{ required: "Please select a country" }}
             render={({
-              field: { onChange, onBlur, value, name, ref },
+              field: { onChange, value, name },
               fieldState: { error },
             }) => (
               <FormControl isRequired isInvalid={Boolean(error)} mr={2}>
-                <FormLabel color="grey.400"> Country </FormLabel>
+                <FormLabel color="grey.400">Country</FormLabel>
                 <Select
                   name={name}
-                  ref={ref}
                   onChange={onChange}
-                  onBlur={onBlur}
                   value={value}
                   options={countryOptions}
                   placeholder=""
-                  chakraStyles={{
-                    dropdownIndicator: (provided) => ({
-                      ...provided,
-                      bg: "transparent",
-                      px: 2,
-                      cursor: "inherit",
-                    }),
-                    indicatorSeparator: (provided) => ({
-                      ...provided,
-                      display: "none",
-                    }),
-                    placeholder: (provided) => ({
-                      ...provided,
-                      color: "grey.300",
-                    }),
-                  }}
+                  useBasicStyles
                 />
                 <FormErrorMessage>{error?.message}</FormErrorMessage>
               </FormControl>
@@ -222,13 +140,13 @@ const CreateAssessmentPage = (): React.ReactElement => {
           />
 
           <FormControl isRequired isInvalid={Boolean(errors.region)}>
-            <FormLabel color="grey.400"> Region </FormLabel>
+            <FormLabel color="grey.400">Region</FormLabel>
             <Input
               {...register("region", { required: "Please enter a region" })}
             />
             <FormErrorMessage> {errors.region?.message} </FormErrorMessage>
           </FormControl>
-        </Stack>
+        </HStack>
       </FormControl>
 
       <Button onClick={handleSubmit(onSubmit, onError)}>Submit</Button>
